backend/pyproject.toml
[tool.poetry]
name = "petcare-backend"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
djangorestframework = "^3.15"
django-cors-headers = "^4.4"
python-dotenv = "^1.0"
psycopg2-binary = "^2.9"
